+++ lib/admin_screens/edit_profile_Admin.dart
@@ -5,7 +5,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart' as firebase_storage;
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
-import 'package:hostelite/alerts_admin.dart';
+import 'package:hostelite/admin_screens/alerts_admin.dart';
 import 'package:hostelite/exit-recordsAdmin.dart';
 import 'package:hostelite/home_screen_Admin.dart';
 import 'package:image_picker/image_picker.dart';
@@ -23,11 +23,10 @@ firebase_storage.FirebaseStorage storage =
 
 firebase_storage.Reference ref = storage.ref().child('dps');
 
-class _EditProfileAdminState extends State<EditProfileAdmin>{
-
+class _EditProfileAdminState extends State<EditProfileAdmin> {
   var imageUrl = 'str';
   File _pickedImage;
-  var picsdb =  FirebaseFirestore.instance
+  var picsdb = FirebaseFirestore.instance
       .collection('displayPics')
       .doc(FirebaseAuth.instance.currentUser.uid);
   String ImgUrl = "";
@@ -42,8 +41,9 @@ class _EditProfileAdminState extends State<EditProfileAdmin>{
       });
     }
   }
-FirebaseAuth _auth = FirebaseAuth.instance;
-FirebaseFirestore db = FirebaseFirestore.instance;
+
+  FirebaseAuth _auth = FirebaseAuth.instance;
+  FirebaseFirestore db = FirebaseFirestore.instance;
 
   final TextEditingController username = TextEditingController();
   final TextEditingController mobileNumber = TextEditingController();
@@ -52,48 +52,44 @@ FirebaseFirestore db = FirebaseFirestore.instance;
     FirebaseFirestore.instance
         .collection("adminUsers")
         .doc(FirebaseAuth.instance.currentUser.uid)
-    .collection("profile")
+        .collection("profile")
         .doc(FirebaseAuth.instance.currentUser.uid)
         .update({
-      "username" : username.text,
-      "mobileNumber" : mobileNumber.text,
-      "emailAddress" : email.text,
+      "username": username.text,
+      "mobileNumber": mobileNumber.text,
+      "emailAddress": email.text,
     });
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-
       body: SafeArea(
         child: SingleChildScrollView(
           child: Container(
             child: Padding(
               padding: EdgeInsets.fromLTRB(30, 50, 15, 15),
               child: Column(children: [
-                Row(children: [IconButton(
-                    icon: Icon( Icons.arrow_back),
-                  onPressed: () => Navigator.of(context).pop(),),
-                     SizedBox(width: 50,),
-                     Text(
-                         'Edit Profile',
-                           style:TextStyle(
-                             fontSize: 24,
-                             color: Color(0xff747475)
-                           )
-                     ),
-                  
-
-
-                ]
+                Row(children: [
+                  IconButton(
+                    icon: Icon(Icons.arrow_back),
+                    onPressed: () => Navigator.of(context).pop(),
+                  ),
+                  SizedBox(
+                    width: 50,
+                  ),
+                  Text('Edit Profile',
+                      style: TextStyle(fontSize: 24, color: Color(0xff747475))),
+                ]),
+                SizedBox(
+                  height: 35,
                 ),
-                SizedBox(height: 35,),
                 GestureDetector(
                   onTap: () async {
                     firebase_storage.UploadTask uploadedImg = ref
                         .child(
-                        DateTime.now().microsecondsSinceEpoch.toString() +
-                            '.png')
+                            DateTime.now().microsecondsSinceEpoch.toString() +
+                                '.png')
                         .putFile(img);
                     await uploadedImg.whenComplete(() => null);
 
@@ -102,91 +98,86 @@ FirebaseFirestore db = FirebaseFirestore.instance;
                     await ref.getDownloadURL().then((value) {
                       url = value;
                     });
-                    db.collection('dps').doc(_auth.currentUser.uid).set({
-                      "dpUrl":url
-                    });
-
-                      },
+                    db
+                        .collection('dps')
+                        .doc(_auth.currentUser.uid)
+                        .set({"dpUrl": url});
+                  },
                   child: Container(
                     child: GestureDetector(
-                    onTap: () {
-                    ImagePicker()
-                        .pickImage(source: ImageSource.gallery)
-                        .then((value) async {
-                                _pickedImage = File(value.path);
-                                Reference reference = FirebaseStorage.instance
-                                    .ref()
-                                    .child('images')
-                                    .child('dps')
-                                    .child(
-                                DateTime.now().microsecondsSinceEpoch.toString() +
-                                '.jpg');
-                                UploadTask task = reference.putFile(_pickedImage);
-                                task.whenComplete(() {
-                                reference.getDownloadURL().then((url) {
-                                setState(() {
+                      onTap: () {
+                        ImagePicker()
+                            .pickImage(source: ImageSource.gallery)
+                            .then((value) async {
+                          _pickedImage = File(value.path);
+                          Reference reference = FirebaseStorage.instance
+                              .ref()
+                              .child('images')
+                              .child('dps')
+                              .child(DateTime.now()
+                                      .microsecondsSinceEpoch
+                                      .toString() +
+                                  '.jpg');
+                          UploadTask task = reference.putFile(_pickedImage);
+                          task.whenComplete(() {
+                            reference.getDownloadURL().then((url) {
+                              setState(() {
                                 ImgUrl = url;
-                                });
-                                print(url);
+                              });
+                              print(url);
 
-                                  picsdb.set({
+                              picsdb.set({
                                 'dpUrl': url,
                                 'userUid': FirebaseAuth.instance.currentUser.uid
-                                  });
-                                });
-                                });
-                    });
-                    },
+                              });
+                            });
+                          });
+                        });
+                      },
                       child: StreamBuilder<QuerySnapshot>(
-                          stream:  FirebaseFirestore.instance
+                          stream: FirebaseFirestore.instance
                               .collection('displayPics')
-                              .where("userUid",isEqualTo:FirebaseAuth.instance.currentUser.uid).snapshots(),
+                              .where("userUid",
+                                  isEqualTo:
+                                      FirebaseAuth.instance.currentUser.uid)
+                              .snapshots(),
                           builder: (context, snapshot) {
-
-                            String dataUrl  = snapshot.data.docs[0]["dpUrl"];
-                            return !snapshot.hasData ? CircularProgressIndicator() : CircleAvatar(
-                              radius: 85,
-                              backgroundColor: Colors.orange[100],
-                              backgroundImage: dataUrl != " "
-                                  ? NetworkImage(dataUrl)
-                                  : AssetImage('assets/nodppic.jfif'),
-                            );
-                          }
-                      ),
+                            String dataUrl = snapshot.data.docs[0]["dpUrl"];
+                            return !snapshot.hasData
+                                ? CircularProgressIndicator()
+                                : CircleAvatar(
+                                    radius: 85,
+                                    backgroundColor: Colors.orange[100],
+                                    backgroundImage: dataUrl != " "
+                                        ? NetworkImage(dataUrl)
+                                        : AssetImage('assets/nodppic.jfif'),
+                                  );
+                          }),
                     ),
                   ),
                 ),
-                SizedBox(height: 35,),
+                SizedBox(
+                  height: 35,
+                ),
                 TextFormField(
                   controller: username,
                   decoration: InputDecoration(
-
                     hintText: _auth.currentUser.displayName,
                     labelText: 'Name',
                     fillColor: Colors.white,
                     filled: true,
                     enabledBorder: OutlineInputBorder(
                         borderRadius: BorderRadius.circular(12.0),
-                        borderSide: BorderSide(
-
-                            color: Colors.grey,
-                            width: 1.0
-                        )
-                    ),
-
+                        borderSide: BorderSide(color: Colors.grey, width: 1.0)),
                     focusedBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                            color: Colors.cyan,
-                            width: 1.0
-                        )
-                    ),
+                        borderSide: BorderSide(color: Colors.cyan, width: 1.0)),
                   ),
                 ),
-                SizedBox(height: 28,),
-
+                SizedBox(
+                  height: 28,
+                ),
                 TextFormField(
                   controller: email,
-
                   decoration: InputDecoration(
                     hintText: _auth.currentUser.email,
                     labelText: 'Email',
@@ -194,24 +185,14 @@ FirebaseFirestore db = FirebaseFirestore.instance;
                     filled: true,
                     enabledBorder: OutlineInputBorder(
                         borderRadius: BorderRadius.circular(12.0),
-                        borderSide: BorderSide(
-
-                            color: Colors.grey,
-                            width: 1.0
-                        )
-                    ),
-
+                        borderSide: BorderSide(color: Colors.grey, width: 1.0)),
                     focusedBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                            color: Colors.cyan,
-                            width: 1.0
-                        )
-                    ),
+                        borderSide: BorderSide(color: Colors.cyan, width: 1.0)),
                   ),
                 ),
-
-                SizedBox(height: 28,),
-
+                SizedBox(
+                  height: 28,
+                ),
                 TextFormField(
                   controller: mobileNumber,
                   decoration: InputDecoration(
@@ -221,55 +202,46 @@ FirebaseFirestore db = FirebaseFirestore.instance;
                     filled: true,
                     enabledBorder: OutlineInputBorder(
                         borderRadius: BorderRadius.circular(12.0),
-                        borderSide: BorderSide(
-
-                            color: Colors.grey,
-                            width: 1.0
-                        )
-                    ),
-
+                        borderSide: BorderSide(color: Colors.grey, width: 1.0)),
                     focusedBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                            color: Colors.cyan,
-                            width: 1.0
-                        )
-                    ),
+                        borderSide: BorderSide(color: Colors.cyan, width: 1.0)),
                   ),
                 ),
-                SizedBox(height:50),
-                
+                SizedBox(height: 50),
                 Container(
                   width: 130,
                   height: 50,
                   child: MaterialButton(
-                    child: Text('Save',
-                      style: TextStyle(color: Colors.white, fontSize: 15),
-                    ),
-                    color: Colors.purple,
-                    minWidth: 100,
-                    shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10.0)),
-                    onPressed: () async {
-
-                        if (username.text.isEmpty || email.text.isEmpty || mobileNumber.text.isEmpty){
-                  showDialog(
-                        context: context,
-                        builder: (BuildContext context) {
-                          return AlertDialog(
-                            title: Text("Error updating profile"),
-                            content: Text("Please fill all fields"),
-                            actions: [
-                              FlatButton(
-                                child: Text("Ok"),
-                                onPressed: () {
-                                  Navigator.of(context).pop();
-                                },
-                              )
-                            ],
-                          );
-
-                        });
-                  return;
-                      }
+                      child: Text(
+                        'Save',
+                        style: TextStyle(color: Colors.white, fontSize: 15),
+                      ),
+                      color: Colors.purple,
+                      minWidth: 100,
+                      shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(10.0)),
+                      onPressed: () async {
+                        if (username.text.isEmpty ||
+                            email.text.isEmpty ||
+                            mobileNumber.text.isEmpty) {
+                          showDialog(
+                              context: context,
+                              builder: (BuildContext context) {
+                                return AlertDialog(
+                                  title: Text("Error updating profile"),
+                                  content: Text("Please fill all fields"),
+                                  actions: [
+                                    FlatButton(
+                                      child: Text("Ok"),
+                                      onPressed: () {
+                                        Navigator.of(context).pop();
+                                      },
+                                    )
+                                  ],
+                                );
+                              });
+                          return;
+                        }
 
                         buildUpdateProfile();
                         print("1.-------------------");
@@ -279,30 +251,21 @@ FirebaseFirestore db = FirebaseFirestore.instance;
                         firebaseUser.updateDisplayName(username.text);
                         await firebaseUser.updateEmail(email.text);
                         print("2.-------------------");
-                          Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (context) => HomeScreenAdmin() ));
-                        }
-
-
-
-                  ),
+                        Navigator.of(context).pushReplacement(MaterialPageRoute(
+                            builder: (context) => HomeScreenAdmin()));
+                      }),
                 ),
-
-                
-
-
               ]),
             ),
           ),
         ),
-
       ),
-      bottomNavigationBar:  Container(
+      bottomNavigationBar: Container(
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(15),
           border: Border.all(
             color: Colors.grey[300],
           ),
-
         ),
         height: 45,
         width: 380,
@@ -313,33 +276,13 @@ FirebaseFirestore db = FirebaseFirestore.instance;
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return HomeScreenAdmin( );
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return HomeScreenAdmin();
+                  }),
                 );
               },
               child: Icon(
-
                 Icons.home_filled,
-              ),
-            ),
-            Spacer(),
-            //SizedBox(width: 10),
-            MaterialButton(
-              onPressed: () {
-
-                Navigator.push(context,
-                  MaterialPageRoute( builder: (context) {
-                    return ExitListAdmin();
-                  }
-                  ),
-                );
-              },
-              child: Icon(
-                Icons.graphic_eq,
-
               ),
             ),
             Spacer(),
@@ -348,32 +291,40 @@ FirebaseFirestore db = FirebaseFirestore.instance;
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return Alerts();
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return ExitListAdmin();
+                  }),
                 );
               },
               child: Icon(
-                Icons.add_alert,
-
+                Icons.graphic_eq,
               ),
             ),
             Spacer(),
             //SizedBox(width: 10),
             MaterialButton(
               onPressed: () {
-
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) {
+                    return Alerts();
+                  }),
+                );
               },
               child: Icon(
-                Icons.person,
-                color:Color(0xffF989E7),
+                Icons.add_alert,
               ),
             ),
             Spacer(),
-
-
+            //SizedBox(width: 10),
+            MaterialButton(
+              onPressed: () {},
+              child: Icon(
+                Icons.person,
+                color: Color(0xffF989E7),
+              ),
+            ),
+            Spacer(),
           ],
         ),
       ),
