+++ lib/entry-recordsAdmin.dart
@@ -1,11 +1,11 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import "package:flutter/material.dart";
-import 'package:hostelite/alerts_admin.dart';
-import 'package:hostelite/edit_profile_Admin.dart';
+import 'package:hostelite/admin_screens/alerts_admin.dart';
+import 'package:hostelite/admin_screens/edit_profile_Admin.dart';
 import 'package:hostelite/exit-recordsAdmin.dart';
 import 'package:hostelite/home_screen_Admin.dart';
-import 'package:hostelite/home_screen_Student.dart';
+import 'package:hostelite/student_screens/home_screen_Student.dart';
 import 'package:hostelite/rejected_complaints.dart';
 import 'package:hostelite/shared_files/decoration.dart';
 
@@ -17,43 +17,33 @@ class EntryListAdmin extends StatefulWidget {
 }
 
 class _EntryListAdminState extends State<EntryListAdmin> {
-
   var db = FirebaseFirestore.instance;
   List entries = [];
-
-
 
   @override
   void initState() {
     super.initState();
-    db
-        .collection('Entries')
-        .get()
-        .then((value) {
+    db.collection('Entries').get().then((value) {
       setState(() {
         entries = value.docs;
       });
-      entries.sort((b,a) => a["time"].compareTo(b["time"]));
+      entries.sort((b, a) => a["time"].compareTo(b["time"]));
     });
   }
-
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: PreferredSize(
-
         preferredSize: Size.fromHeight(100),
         child: AppBar(
-          leading: BackButton(
-              color: Colors.black
-          ),
-          title: Text('Entry-Exit Reports',
+          leading: BackButton(color: Colors.black),
+          title: Text(
+            'Entry-Exit Reports',
             style: TextStyle(
               color: Color(0xff4E4E4E),
-            ),),
-
-
+            ),
+          ),
           backgroundColor: Color(0xffFE96FA),
         ),
       ),
@@ -63,59 +53,54 @@ class _EntryListAdminState extends State<EntryListAdmin> {
             Padding(
               padding: const EdgeInsets.all(16.0),
               child: Card(
-                  child:Padding(
-
-                    padding: const EdgeInsets.all(8.0),
-                    child: Row(
-                      children: [
-                        Padding(
-                          padding: const EdgeInsets.only(left:50.0),
-                          child: Container(
-                            decoration: BoxDecoration(
-                                border: Border(
-
-                                  bottom: BorderSide(
-
-                                      width: 5.0, color: Color(0xffFE96FA)),
-                                )
-                            ),
-                            child: Text('Entry',style: TextStyle(
-                                fontSize: 18
-                            ),),
-                          ),
+                  child: Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: Row(
+                  children: [
+                    Padding(
+                      padding: const EdgeInsets.only(left: 50.0),
+                      child: Container(
+                        decoration: BoxDecoration(
+                            border: Border(
+                          bottom:
+                              BorderSide(width: 5.0, color: Color(0xffFE96FA)),
+                        )),
+                        child: Text(
+                          'Entry',
+                          style: TextStyle(fontSize: 18),
                         ),
-                        Spacer(),
-                        Padding(
-                          padding: const EdgeInsets.only(right: 50),
-
-                          child: GestureDetector(
-                            onTap: (){
-                              Navigator.push(
-                                context,
-                                MaterialPageRoute(builder: (context) {
-                                  return ExitListAdmin();
-                                }),
-                              );
-                            },
-                            child: Container(
-
-                              child: Text('Exit',style: TextStyle(
-                                  fontSize: 18
-                              ),),
-                            ),
-                          ),
-                        ),
-
-                      ],
+                      ),
                     ),
-                  )
-              ),
+                    Spacer(),
+                    Padding(
+                      padding: const EdgeInsets.only(right: 50),
+                      child: GestureDetector(
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(builder: (context) {
+                              return ExitListAdmin();
+                            }),
+                          );
+                        },
+                        child: Container(
+                          child: Text(
+                            'Exit',
+                            style: TextStyle(fontSize: 18),
+                          ),
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              )),
             ),
 
             SingleChildScrollView(
               scrollDirection: Axis.horizontal,
               child: Theme(
-                data: Theme.of(context).copyWith(dividerColor: Colors.blueAccent),
+                data:
+                    Theme.of(context).copyWith(dividerColor: Colors.blueAccent),
                 child: DataTable(
                   sortAscending: true,
                   columns: [
@@ -126,22 +111,28 @@ class _EntryListAdminState extends State<EntryListAdmin> {
                     DataColumn(label: Text('Location')),
                     DataColumn(label: Text('Hostel')),
                     DataColumn(label: Text('Room No.')),
-
                   ],
-                  rows:
-                  entries.map((element) => DataRow(
-                    cells: <DataCell>[
-                      DataCell(Text(element["time"].toDate().toString().substring(0,11))), //Extracting from Map element the value
-                      DataCell(Text(element["time"].toDate().toString().substring(11,19))), //Extracting from Map element the value
-                      DataCell(Text(element["name"])),
-                      DataCell(Text(element["rollNumber"])),
-                      DataCell(Text(element["location"])),
-                      DataCell(Text(element["hostel"])),
-                      DataCell(Text(element["roomNumber"])),
-
-                    ],
-                  )
-                  ).toList(),
+                  rows: entries
+                      .map((element) => DataRow(
+                            cells: <DataCell>[
+                              DataCell(Text(element["time"]
+                                  .toDate()
+                                  .toString()
+                                  .substring(0,
+                                      11))), //Extracting from Map element the value
+                              DataCell(Text(element["time"]
+                                  .toDate()
+                                  .toString()
+                                  .substring(11,
+                                      19))), //Extracting from Map element the value
+                              DataCell(Text(element["name"])),
+                              DataCell(Text(element["rollNumber"])),
+                              DataCell(Text(element["location"])),
+                              DataCell(Text(element["hostel"])),
+                              DataCell(Text(element["roomNumber"])),
+                            ],
+                          ))
+                      .toList(),
                 ),
               ),
             ),
@@ -217,15 +208,12 @@ class _EntryListAdminState extends State<EntryListAdmin> {
       //   ),
       //   SizedBox(height: 10.0,),
 
-
-
-      bottomNavigationBar:  Container(
+      bottomNavigationBar: Container(
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(15),
           border: Border.all(
             color: Colors.grey[300],
           ),
-
         ),
         height: 45,
         width: 380,
@@ -236,15 +224,12 @@ class _EntryListAdminState extends State<EntryListAdmin> {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return HomeScreenAdmin( );
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return HomeScreenAdmin();
+                  }),
                 );
               },
               child: Icon(
-
                 Icons.home_filled,
               ),
             ),
@@ -254,7 +239,7 @@ class _EntryListAdminState extends State<EntryListAdmin> {
               onPressed: () {},
               child: Icon(
                 Icons.graphic_eq,
-                color:Color(0xffF989E7),
+                color: Color(0xffF989E7),
               ),
             ),
             Spacer(),
@@ -263,11 +248,9 @@ class _EntryListAdminState extends State<EntryListAdmin> {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return Alerts();
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return Alerts();
+                  }),
                 );
               },
               child: Icon(
@@ -280,11 +263,9 @@ class _EntryListAdminState extends State<EntryListAdmin> {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return EditProfileAdmin();
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return EditProfileAdmin();
+                  }),
                 );
               },
               child: Icon(
@@ -292,14 +273,9 @@ class _EntryListAdminState extends State<EntryListAdmin> {
               ),
             ),
             Spacer(),
-
-
           ],
         ),
       ),
-
-
     );
   }
 }
-
