+++ lib/exit-recordsAdmin.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'package:hostelite/alerts_admin.dart';
-import 'package:hostelite/edit_profile_Admin.dart';
+import 'package:hostelite/admin_screens/alerts_admin.dart';
+import 'package:hostelite/admin_screens/edit_profile_Admin.dart';
 import 'package:hostelite/entry-recordsAdmin.dart';
 import 'package:hostelite/home_screen_Admin.dart';
 
@@ -13,148 +13,131 @@ class ExitListAdmin extends StatefulWidget {
 }
 
 class _ExitListAdminState extends State<ExitListAdmin> {
-
   var db = FirebaseFirestore.instance;
   List exits = [];
-
-
 
   @override
   void initState() {
     super.initState();
-    db
-        .collection('Exits')
-        .get()
-        .then((value) {
+    db.collection('Exits').get().then((value) {
       setState(() {
         exits = value.docs;
       });
 
-      exits.sort((b,a) => a["time"].compareTo(b["time"]));
-
+      exits.sort((b, a) => a["time"].compareTo(b["time"]));
     });
   }
-
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: PreferredSize(
-
         preferredSize: Size.fromHeight(100),
         child: AppBar(
-          leading: BackButton(
-              color: Colors.black
-          ),
-          title: Text('Entry-Exit Reports',
+          leading: BackButton(color: Colors.black),
+          title: Text(
+            'Entry-Exit Reports',
             style: TextStyle(
               color: Color(0xff4E4E4E),
-            ),),
-
-
+            ),
+          ),
           backgroundColor: Color(0xffFE96FA),
         ),
       ),
       body: SingleChildScrollView(
-        child: Column(
-            children: [
-            Padding(
+        child: Column(children: [
+          Padding(
             padding: const EdgeInsets.all(16.0),
             child: Card(
-                child:Padding(
-
-                  padding: const EdgeInsets.all(8.0),
-                  child: Row(
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.only(left:50.0),
-                        child: GestureDetector(
-                          onTap: (){
-                            Navigator.push(
-                              context,
-                              MaterialPageRoute(builder: (context) {
-                                return EntryListAdmin();
-                              }),
-                            );
-                          },
-                          child: Container(
-
-                            child: Text('Entry',style: TextStyle(
-                                fontSize: 18
-                            ),),
-                          ),
+                child: Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: Row(
+                children: [
+                  Padding(
+                    padding: const EdgeInsets.only(left: 50.0),
+                    child: GestureDetector(
+                      onTap: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(builder: (context) {
+                            return EntryListAdmin();
+                          }),
+                        );
+                      },
+                      child: Container(
+                        child: Text(
+                          'Entry',
+                          style: TextStyle(fontSize: 18),
                         ),
                       ),
-                      Spacer(),
-                      Padding(
-                        padding: const EdgeInsets.only(right: 50),
-
-                        child: Container(
-                          decoration: BoxDecoration(
-                              border: Border(
-
-                                bottom: BorderSide(
-
-                                    width: 5.0, color: Color(0xffFE96FA)),
-                              )
-                          ),
-
-                          child: Text('Exit',style: TextStyle(
-                              fontSize: 18
-                          ),),
-                        ),
-                      ),
-
-                    ],
+                    ),
                   ),
-                )
+                  Spacer(),
+                  Padding(
+                    padding: const EdgeInsets.only(right: 50),
+                    child: Container(
+                      decoration: BoxDecoration(
+                          border: Border(
+                        bottom:
+                            BorderSide(width: 5.0, color: Color(0xffFE96FA)),
+                      )),
+                      child: Text(
+                        'Exit',
+                        style: TextStyle(fontSize: 18),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            )),
+          ),
+          SingleChildScrollView(
+            scrollDirection: Axis.horizontal,
+            child: Theme(
+              data: Theme.of(context).copyWith(dividerColor: Colors.blueAccent),
+              child: DataTable(
+                sortAscending: true,
+                columns: [
+                  DataColumn(label: Text('Date')),
+                  DataColumn(label: Text('Time')),
+                  DataColumn(label: Text('User email')),
+                  DataColumn(label: Text('Roll No.')),
+                  DataColumn(label: Text('purpose')),
+                  DataColumn(label: Text('Hostel')),
+                  DataColumn(label: Text('Room No.')),
+                ],
+                rows: exits
+                    .map((element) => DataRow(
+                          cells: <DataCell>[
+                            DataCell(Text(element["time"]
+                                .toDate()
+                                .toString()
+                                .substring(0,
+                                    11))), //Extracting from Map element the value
+                            DataCell(Text(element["time"]
+                                .toDate()
+                                .toString()
+                                .substring(11,
+                                    19))), //Extracting from Map element the value
+                            DataCell(Text(element["name"])),
+                            DataCell(Text(element["rollNumber"])),
+                            DataCell(Text(element["purpose"])),
+                            DataCell(Text(element["hostelName"])),
+                            DataCell(Text(element["roomNumber"])),
+                          ],
+                        ))
+                    .toList(),
+              ),
             ),
           ),
-
-              SingleChildScrollView(
-                scrollDirection: Axis.horizontal,
-                child: Theme(
-                  data: Theme.of(context).copyWith(dividerColor: Colors.blueAccent),
-                  child: DataTable(
-
-                    sortAscending: true,
-                    columns: [
-                      DataColumn(label: Text('Date')),
-                      DataColumn(label: Text('Time')),
-                      DataColumn(label: Text('User email')),
-                      DataColumn(label: Text('Roll No.')),
-                      DataColumn(label: Text('purpose')),
-                      DataColumn(label: Text('Hostel')),
-                      DataColumn(label: Text('Room No.')),
-
-                    ],
-                    rows:
-                    exits.map((element) => DataRow(
-                      cells: <DataCell>[
-                        DataCell(Text(element["time"].toDate().toString().substring(0,11))), //Extracting from Map element the value
-                        DataCell(Text(element["time"].toDate().toString().substring(11,19))), //Extracting from Map element the value
-                        DataCell(Text(element["name"])),
-                        DataCell(Text(element["rollNumber"])),
-                        DataCell(Text(element["purpose"])),
-                        DataCell(Text(element["hostelName"])),
-                        DataCell(Text(element["roomNumber"])),
-
-                      ],
-                    )
-                    ).toList(),
-                  ),
-                ),
-              ),
-
-          ]),
+        ]),
       ),
-      bottomNavigationBar:  Container(
+      bottomNavigationBar: Container(
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(15),
           border: Border.all(
             color: Colors.grey[300],
           ),
-
         ),
         height: 45,
         width: 380,
@@ -165,15 +148,12 @@ class _ExitListAdminState extends State<ExitListAdmin> {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return HomeScreenAdmin( );
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return HomeScreenAdmin();
+                  }),
                 );
               },
               child: Icon(
-
                 Icons.home_filled,
               ),
             ),
@@ -183,7 +163,7 @@ class _ExitListAdminState extends State<ExitListAdmin> {
               onPressed: () {},
               child: Icon(
                 Icons.graphic_eq,
-                color:Color(0xffF989E7),
+                color: Color(0xffF989E7),
               ),
             ),
             Spacer(),
@@ -192,11 +172,9 @@ class _ExitListAdminState extends State<ExitListAdmin> {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return Alerts();
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return Alerts();
+                  }),
                 );
               },
               child: Icon(
@@ -209,11 +187,9 @@ class _ExitListAdminState extends State<ExitListAdmin> {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return EditProfileAdmin();
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return EditProfileAdmin();
+                  }),
                 );
               },
               child: Icon(
@@ -221,12 +197,9 @@ class _ExitListAdminState extends State<ExitListAdmin> {
               ),
             ),
             Spacer(),
-
-
           ],
         ),
       ),
     );
-
   }
 }
