+++ lib/rejected_complaints.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:hostelite/alerts_admin.dart';
-import 'package:hostelite/complaintAdmin.dart';
-import 'package:hostelite/edit_profile_Admin.dart';
+import 'package:hostelite/admin_screens/alerts_admin.dart';
+import 'package:hostelite/admin_screens/complaintAdmin.dart';
+import 'package:hostelite/admin_screens/edit_profile_Admin.dart';
 import 'package:hostelite/home_screen_Admin.dart';
 import 'package:hostelite/pendingcomplaints_admin.dart';
 
@@ -13,27 +13,31 @@ class RejectedComplaints extends StatefulWidget {
 }
 
 class _RejectedComplaintsState extends State<RejectedComplaints> {
-
   String dropdownvalue = "Category";
-  var items = ['Category','Infrastructure','Sanitation','Water','Electricity','Wifi','Others'];
+  var items = [
+    'Category',
+    'Infrastructure',
+    'Sanitation',
+    'Water',
+    'Electricity',
+    'Wifi',
+    'Others'
+  ];
   final now = DateTime.now();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: PreferredSize(
-
         preferredSize: Size.fromHeight(100),
         child: AppBar(
-          leading: BackButton(
-              color: Colors.black
-          ),
-          title: Text('View Complaints',
+          leading: BackButton(color: Colors.black),
+          title: Text(
+            'View Complaints',
             style: TextStyle(
               color: Color(0xff4E4E4E),
-            ),),
-
-
+            ),
+          ),
           backgroundColor: Color(0xffFE96FA),
         ),
       ),
@@ -48,11 +52,11 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
                     padding: EdgeInsets.all(5),
                     child: Row(
                       children: [
-                        Text('Status : ',
+                        Text(
+                          'Status : ',
                           style: TextStyle(
-                              fontSize: 15,
-                              fontWeight: FontWeight.w600
-                          ),),
+                              fontSize: 15, fontWeight: FontWeight.w600),
+                        ),
                         Spacer(),
                         TextButton(
                           style: TextButton.styleFrom(
@@ -61,60 +65,52 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
                                 fontWeight: FontWeight.w700,
                                 color: Color(0xff4BB227)),
                           ),
-                          onPressed: (){
+                          onPressed: () {
                             Navigator.push(
                               context,
-                              MaterialPageRoute(
-                                  builder: (context) {
-                                    return ViewComplaintsAdmin();
-                                  }
-                              ),
+                              MaterialPageRoute(builder: (context) {
+                                return ViewComplaintsAdmin();
+                              }),
                             );
                           },
-                          child: const Text('Sorted',
-                            style: TextStyle(
-
-                                color: Color(0xff4BB227)
-                            ),),
+                          child: const Text(
+                            'Sorted',
+                            style: TextStyle(color: Color(0xff4BB227)),
+                          ),
                         ),
                         Spacer(),
                         VerticalDivider(
                           color: Colors.green,
-                          thickness: 2, width: 2,
+                          thickness: 2,
+                          width: 2,
                         ),
                         Spacer(),
                         TextButton(
                           style: TextButton.styleFrom(
-                            primary:  Colors.red,
+                            primary: Colors.red,
                             textStyle: const TextStyle(
                               fontSize: 15,
                               fontWeight: FontWeight.w700,
                             ),
-
                           ),
                           onPressed: () {
                             Navigator.push(
                               context,
-                              MaterialPageRoute(
-                                  builder: (context) {
-                                    return PendingComplaints();
-                                  }
-                              ),
+                              MaterialPageRoute(builder: (context) {
+                                return PendingComplaints();
+                              }),
                             );
-
                           },
-                          child: const Text('Pending',
-
-                            style: TextStyle(
-
-                                color: Color(0xffFBC02D)
-                            ),
+                          child: const Text(
+                            'Pending',
+                            style: TextStyle(color: Color(0xffFBC02D)),
                           ),
                         ),
                         Spacer(),
                         VerticalDivider(
                           color: Colors.lightBlue,
-                          thickness: 2, width: 4,
+                          thickness: 2,
+                          width: 4,
                         ),
                         Spacer(),
                         TextButton(
@@ -125,12 +121,13 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
                             ),
                           ),
                           onPressed: null,
-                          child: const Text('Rejected',
+                          child: const Text(
+                            'Rejected',
                             style: TextStyle(
                                 decoration: TextDecoration.underline,
                                 decorationThickness: 4,
-                                color: Color(0xffFF4646)
-                            ),),
+                                color: Color(0xffFF4646)),
+                          ),
                         ),
                         Spacer(),
                       ],
@@ -138,37 +135,28 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
                   ),
                 ),
               ),
-
-
             ),
             Container(
               decoration: BoxDecoration(
                   borderRadius: BorderRadius.circular(15),
-                  boxShadow:[ BoxShadow(
-
-                      color: Colors.tealAccent.withOpacity(0.1),
-                      spreadRadius: 5
-                  ),
-                  ]
-              ),
+                  boxShadow: [
+                    BoxShadow(
+                        color: Colors.tealAccent.withOpacity(0.1),
+                        spreadRadius: 5),
+                  ]),
               child: Row(
                 children: [
-
                   Spacer(),
                   Padding(
-                    padding:EdgeInsets.fromLTRB(30, 5, 30, 5),
+                    padding: EdgeInsets.fromLTRB(30, 5, 30, 5),
                     child: DropdownButton(
-
                       value: dropdownvalue,
                       icon: Icon(Icons.keyboard_arrow_down),
-                      items:items.map((String items) {
+                      items: items.map((String items) {
                         return DropdownMenuItem(
-                            value: items,
-                            child: Text(items)
-                        );
-                      }
-                      ).toList(),
-                      onChanged: (String newValue){
+                            value: items, child: Text(items));
+                      }).toList(),
+                      onChanged: (String newValue) {
                         setState(() {
                           dropdownvalue = newValue;
                         });
@@ -180,20 +168,18 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
                   //child: DropdownDatePicker(
                   //firstDate: ValidDate(year: now.year - 100, month: 1, day: 1),
                   //lastDate: ValidDate(year: now.year, month: now.month, day: now.month),
-
-
                 ],
               ),
             ),
-            SizedBox(height: 20,),
+            SizedBox(
+              height: 20,
+            ),
             Container(
-
               child: Padding(
-                padding: EdgeInsets.fromLTRB(15,0,15,0),
+                padding: EdgeInsets.fromLTRB(15, 0, 15, 0),
                 child: Card(
                   shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(15)
-                  ),
+                      borderRadius: BorderRadius.circular(15)),
                   child: Padding(
                     padding: EdgeInsets.all(15),
                     child: Column(
@@ -203,71 +189,85 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
                             CircleAvatar(
                               backgroundColor: Colors.yellow,
                             ),
-                            SizedBox(width: 10,),
+                            SizedBox(
+                              width: 10,
+                            ),
                             Column(
                               crossAxisAlignment: CrossAxisAlignment.start,
                               children: [
-                                Text('Ramesh Kumar',
+                                Text(
+                                  'Ramesh Kumar',
                                   style: TextStyle(
-                                      color: Color(0xff565656),
-                                      fontSize: 15
-                                  ),),
-                                Text('Room No. 087',
-                                  style: TextStyle(
-                                      fontSize: 12
-                                  ),),
+                                      color: Color(0xff565656), fontSize: 15),
+                                ),
+                                Text(
+                                  'Room No. 087',
+                                  style: TextStyle(fontSize: 12),
+                                ),
                               ],
                             ),
                             Spacer(),
-                            Text(DateTime.now().day.toString()+'.'+DateTime.now().month.toString()+'.'+DateTime.now().year                     .toString(),
-                              style: TextStyle(
-                                  color: Color(0xff9F9F9F)
-                              ),)
+                            Text(
+                              DateTime.now().day.toString() +
+                                  '.' +
+                                  DateTime.now().month.toString() +
+                                  '.' +
+                                  DateTime.now().year.toString(),
+                              style: TextStyle(color: Color(0xff9F9F9F)),
+                            )
                           ],
                         ),
                         Divider(),
-                        SizedBox(height: 5,),
-                        Image(
-                          image: AssetImage('assets/home_Screen_Student/2720490 1.png'),
+                        SizedBox(
+                          height: 5,
                         ),
-                        SizedBox(height: 5,),
-                        Text('Title of Issue',
-
+                        Image(
+                          image: AssetImage(
+                              'assets/home_Screen_Student/2720490 1.png'),
+                        ),
+                        SizedBox(
+                          height: 5,
+                        ),
+                        Text(
+                          'Title of Issue',
                           style: TextStyle(
                             fontSize: 18,
                             fontWeight: FontWeight.w500,
                             color: Color(0xff1A1919),
                           ),
                         ),
-                        SizedBox(height: 5,),
-                        Text('Sir I went to purchase a book , but while returning it satrted raining . So there was                              no any options left for me then waiting at  the store until the rain stopped.'),
-                        SizedBox(height: 10,),
+                        SizedBox(
+                          height: 5,
+                        ),
+                        Text(
+                            'Sir I went to purchase a book , but while returning it satrted raining . So there was                              no any options left for me then waiting at  the store until the rain stopped.'),
+                        SizedBox(
+                          height: 10,
+                        ),
                         Row(
                           mainAxisAlignment: MainAxisAlignment.end,
                           children: [
                             Text(
                               DateTime.now().toString(),
-                              style: TextStyle(
-                                  color: Color(0xff9F9F9F)
-                              ),)
+                              style: TextStyle(color: Color(0xff9F9F9F)),
+                            )
                           ],
                         )
-
                       ],
                     ),
                   ),
                 ),
               ),
             ),
-            SizedBox(height: 20,),
+            SizedBox(
+              height: 20,
+            ),
             Container(
-
               child: Padding(
-                padding: EdgeInsets.fromLTRB(15,0,15,0),
+                padding: EdgeInsets.fromLTRB(15, 0, 15, 0),
                 child: Card(
                   shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(15)
-                  ),
+                      borderRadius: BorderRadius.circular(15)),
                   child: Padding(
                     padding: EdgeInsets.all(15),
                     child: Column(
@@ -275,76 +275,88 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
                         Row(
                           children: [
                             CircleAvatar(),
-                            SizedBox(width: 10,),
+                            SizedBox(
+                              width: 10,
+                            ),
                             Column(
                               crossAxisAlignment: CrossAxisAlignment.start,
                               children: [
-                                Text('Ramesh Kumar',
+                                Text(
+                                  'Ramesh Kumar',
                                   style: TextStyle(
-                                      color: Color(0xff565656),
-                                      fontSize: 15
-                                  ),),
-                                Text('Room No. 087',
-                                  style: TextStyle(
-                                      fontSize: 12
-                                  ),),
+                                      color: Color(0xff565656), fontSize: 15),
+                                ),
+                                Text(
+                                  'Room No. 087',
+                                  style: TextStyle(fontSize: 12),
+                                ),
                               ],
                             ),
                             Spacer(),
-                            Text(DateTime.now().day.toString()+'.'+DateTime.now().month.toString()+'.'+DateTime.now().year                     .toString(),
-                              style: TextStyle(
-                                  color: Color(0xff9F9F9F)
-                              ),)
+                            Text(
+                              DateTime.now().day.toString() +
+                                  '.' +
+                                  DateTime.now().month.toString() +
+                                  '.' +
+                                  DateTime.now().year.toString(),
+                              style: TextStyle(color: Color(0xff9F9F9F)),
+                            )
                           ],
                         ),
                         Divider(),
-                        SizedBox(height: 5,),
-                        Image(
-                          image: AssetImage('assets/home_Screen_Student/2720490 1.png'),
+                        SizedBox(
+                          height: 5,
                         ),
-                        SizedBox(height: 5,),
-                        Text('Title of Issue',
-
+                        Image(
+                          image: AssetImage(
+                              'assets/home_Screen_Student/2720490 1.png'),
+                        ),
+                        SizedBox(
+                          height: 5,
+                        ),
+                        Text(
+                          'Title of Issue',
                           style: TextStyle(
                             fontSize: 18,
                             fontWeight: FontWeight.w500,
                             color: Color(0xff1A1919),
                           ),
                         ),
-                        SizedBox(height: 5,),
-                        Text('Sir I went to purchase a book , but while returning it satrted raining . So there was                              no any options left for me then waiting at  the store until the rain stopped.'),
-                        SizedBox(height: 10,),
+                        SizedBox(
+                          height: 5,
+                        ),
+                        Text(
+                            'Sir I went to purchase a book , but while returning it satrted raining . So there was                              no any options left for me then waiting at  the store until the rain stopped.'),
+                        SizedBox(
+                          height: 10,
+                        ),
                         Row(
                           mainAxisAlignment: MainAxisAlignment.end,
                           children: [
                             Text(
                               DateTime.now().toString(),
-                              style: TextStyle(
-                                  color: Color(0xff9F9F9F)
-                              ),)
+                              style: TextStyle(color: Color(0xff9F9F9F)),
+                            )
                           ],
                         )
-
                       ],
                     ),
                   ),
                 ),
               ),
             ),
-            SizedBox(height: 20,),
-
-
-
+            SizedBox(
+              height: 20,
+            ),
           ],
         ),
       ),
-      bottomNavigationBar:  Container(
+      bottomNavigationBar: Container(
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(15),
           border: Border.all(
             color: Colors.grey[300],
           ),
-
         ),
         height: 45,
         width: 380,
@@ -355,15 +367,12 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return HomeScreenAdmin( );
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return HomeScreenAdmin();
+                  }),
                 );
               },
               child: Icon(
-
                 Icons.home_filled,
               ),
             ),
@@ -381,11 +390,9 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return Alerts();
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return Alerts();
+                  }),
                 );
               },
               child: Icon(
@@ -398,11 +405,9 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
               onPressed: () {
                 Navigator.push(
                   context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return EditProfileAdmin();
-                      }
-                  ),
+                  MaterialPageRoute(builder: (context) {
+                    return EditProfileAdmin();
+                  }),
                 );
               },
               child: Icon(
@@ -410,12 +415,9 @@ class _RejectedComplaintsState extends State<RejectedComplaints> {
               ),
             ),
             Spacer(),
-
-
           ],
         ),
       ),
-
     );
   }
 }
