+++ lib/student_screens/complaintStudent.dart
@@ -5,7 +5,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:hostelite/home_screen_Student.dart';
+import 'package:hostelite/student_screens/home_screen_Student.dart';
 import 'package:image_picker/image_picker.dart';
 
 class StudentComplaint extends StatefulWidget {
