+++ lib/student_screens/exit_entryStudent.dart
@@ -3,15 +3,16 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import "package:flutter/material.dart";
 import 'package:hostelite/home_screen_Admin.dart';
-import 'package:hostelite/home_screen_Student.dart';
+import 'package:hostelite/student_screens/home_screen_Student.dart';
 import 'package:hostelite/models/user_model.dart';
 import 'package:hostelite/shared_files/decoration.dart';
+import 'package:geolocator/geolocator.dart';
 
-class MarkingExit extends StatefulWidget {
-  const MarkingExit({Key key}) : super(key: key);
+class MarkingEntry extends StatefulWidget {
+  const MarkingEntry({Key key}) : super(key: key);
 
   @override
-  _MarkingExitState createState() => _MarkingExitState();
+  _MarkingEntryState createState() => _MarkingEntryState();
 }
 
 Dialog leadDialog = Dialog(
@@ -66,32 +67,89 @@ Dialog leadDialog = Dialog(
   ),
 );
 
-class _MarkingExitState extends State<MarkingExit> {
+Dialog leadDialogLate = Dialog(
+  child: Container(
+    height: 300,
+    width: 360,
+    child: Column(
+      children: <Widget>[
+        Center(
+          child: Image(
+            image: AssetImage('assets/create_account_page/Vector.png'),
+            height: 150,
+            width: 150,
+          ),
+        ),
+        Text(
+          'You are not',
+          style: TextStyle(
+            fontSize: 30,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        Text(
+          'at hostel',
+          style: TextStyle(
+            fontWeight: FontWeight.bold,
+            fontSize: 30,
+          ),
+        ),
+        Container(
+          width: 150.0,
+          height: 40.0,
+          child: MaterialButton(
+            child: Text(
+              'Done',
+              style: TextStyle(color: Colors.white, fontSize: 17.0),
+            ),
+            color: Colors.pinkAccent[100],
+            minWidth: 100.0,
+            shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(10.0)),
+            onPressed: () {
+              BuildContext context;
+              Navigator.push(context, MaterialPageRoute(builder: (context) {
+                return HomeScreenStudent();
+              }));
+            },
+          ),
+        ),
+      ],
+    ),
+  ),
+);
+
+class _MarkingEntryState extends State<MarkingEntry> {
+  var db = FirebaseFirestore.instance;
+  List alerts = [];
+
   String roomNumber;
   String rollNumber;
-  String purpose;
-  String hostelName;
+  String studentloc;
+  String hostel;
+  int currentTime = DateTime.now().toLocal().hour;
+  var locationMessage = '';
+  Position position;
+
+  void getCurrentLocation() async {
+    position = await Geolocator.getCurrentPosition(
+        desiredAccuracy: LocationAccuracy.high);
+    setState(() {
+      locationMessage =
+          'Longitude: ${position.longitude} \n Latitude: ${position.latitude}';
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        leading: new IconButton(
-          icon: new Icon(Icons.arrow_back_rounded),
-          onPressed: () => Navigator.of(context).pop(),
-        ),
-        title: Text(
-          'Mark Exit',
-          style: TextStyle(
-            fontSize: 20,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        elevation: 10,
+        title: Text('Mark Entry'),
         backgroundColor: Color(0xffFE96FA),
+        elevation: 0,
       ),
       body: Padding(
-        padding: const EdgeInsets.all(20),
+        padding: const EdgeInsets.fromLTRB(20, 30, 20, 0),
         child: Container(
           constraints: BoxConstraints.expand(),
           decoration: BoxDecoration(
@@ -166,86 +224,125 @@ class _MarkingExitState extends State<MarkingExit> {
                       ),
                       onChanged: (value) {
                         setState(() {
-                          hostelName = value.trim();
+                          hostel = value.trim();
                         });
                       }),
                 ),
                 SizedBox(
                   height: 30,
                 ),
-                Container(
-                  child: TextField(
-                      keyboardType: TextInputType.multiline,
-                      maxLines: null,
-                      decoration: InputDecoration(
-                        hintText: 'Purpose of Exit',
-                        filled: true,
-                        fillColor: Color(0xffFFFFFF),
-                        enabledBorder: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12.0),
-                          borderSide:
-                              BorderSide(color: Colors.grey, width: 1.0),
+                Card(
+                  child: Container(
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(15.0),
+                      boxShadow: [
+                        BoxShadow(
+                          color: Colors.white38.withOpacity(0.5),
+                          spreadRadius: 2,
+                          blurRadius: 7,
+                          offset: Offset(1, 3), // changes position of shadow
                         ),
-                        focusedBorder: OutlineInputBorder(
-                            borderSide:
-                                BorderSide(color: Colors.cyan, width: 1.0)),
-                      ),
-                      onChanged: (value) {
-                        setState(() {
-                          purpose = value.trim();
-                        });
-                      }),
+                      ],
+                    ),
+                    child: Column(
+                      children: [
+                        Padding(
+                            padding: EdgeInsets.all(8),
+                            child: Container(
+                              child: Text('Your Location',
+                                  style: TextStyle(
+                                      fontSize: 18,
+                                      color: Colors.green,
+                                      fontWeight: FontWeight.w600)),
+                            )),
+                        Padding(
+                          padding: const EdgeInsets.all(8.0),
+                          child: Text(locationMessage,
+                              style: TextStyle(
+                                fontSize: 16,
+                              )),
+                        ),
+                      ],
+                    ),
+                  ),
                 ),
                 SizedBox(height: 80),
+                MaterialButton(
+                  child: Text(
+                    'Get location',
+                    style: TextStyle(color: Colors.black87, fontSize: 15),
+                  ),
+                  color: Color(0xffFE96FA),
+                  minWidth: 100,
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(10.0)),
+                  onPressed: () {
+                    getCurrentLocation();
+                  },
+                ),
+                SizedBox(height: 40),
                 Container(
                   width: 130,
                   height: 50,
                   child: MaterialButton(
                     child: Text(
-                      'Mark Exit',
-                      style: TextStyle(color: Color(0xff33004A), fontSize: 15),
+                      'Mark Entry',
+                      style: TextStyle(color: Colors.black87, fontSize: 15),
                     ),
                     color: Color(0xffFE96FA),
                     minWidth: 100,
-                    elevation: 10,
                     shape: RoundedRectangleBorder(
                         borderRadius: BorderRadius.circular(10.0)),
                     onPressed: () async {
                       if (roomNumber.isEmpty ||
                           rollNumber.isEmpty ||
-                          hostelName.isEmpty ||
-                          purpose.isEmpty) {
+                          hostel.isEmpty ||
+                          locationMessage.isEmpty) {
                         return SnackBar(
                             content: Text("Please enter all fields"));
                       }
                       FirebaseFirestore.instance
                           .collection("studentUsers")
                           .doc(FirebaseAuth.instance.currentUser.uid)
-                          .collection("exit")
+                          .collection("entry")
                           .add({
-                        "hostelName": hostelName,
+                        // "hostelName" : hostelName,
                         "rollNumber": rollNumber,
                         "roomNumber": roomNumber,
-                        "purpose": purpose,
+                        "hostel": hostel,
                         "userUid": FirebaseAuth.instance.currentUser.uid,
                         "time": DateTime.now().toLocal(),
-                        "name": FirebaseAuth.instance.currentUser.displayName,
+                        "location": locationMessage,
                       });
 
-                      FirebaseFirestore.instance.collection('Exits').add({
-                        "hostelName": hostelName,
+                      FirebaseFirestore.instance.collection('Entries').add({
+                        "name": FirebaseAuth.instance.currentUser.email,
                         "rollNumber": rollNumber,
                         "roomNumber": roomNumber,
-                        "purpose": purpose,
+                        "hostel": hostel,
                         "userUid": FirebaseAuth.instance.currentUser.uid,
                         "time": DateTime.now().toLocal(),
-                        "name": FirebaseAuth.instance.currentUser.displayName,
+                        "location": locationMessage,
                       });
 
-                      showDialog(
-                          context: context,
-                          // ignore: non_constant_identifier_names
-                          builder: (BuildContext) => leadDialog);
+                      if ((currentTime > 8)) {
+                        FirebaseFirestore.instance.collection('alerts').add({
+                          "position": position.toString(),
+                          "name": FirebaseAuth.instance.currentUser.displayName,
+                          "rollNumber": rollNumber,
+                          "roomNumber": roomNumber,
+                          "hostel": hostel,
+                          "userUid": FirebaseAuth.instance.currentUser.uid,
+                          "time": DateTime.now().toLocal(),
+                        });
+                        showDialog(
+                            context: context,
+                            builder: (BuildContext context) => leadDialogLate);
+                      } else {
+                        showDialog(
+                            context: context,
+                            builder: (BuildContext context) => leadDialog);
+                      }
                     },
                   ),
                 ),
@@ -255,6 +352,5 @@ class _MarkingExitState extends State<MarkingExit> {
         ),
       ),
     );
-    ;
   }
 }
