+++ lib/studentsentryrecords.dart
@@ -1,8 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:hostelite/edit_profile_Student.dart';
-import 'package:hostelite/home_screen_Student.dart';
+import 'package:hostelite/student_screens/edit_profile_Student.dart';
+import 'package:hostelite/student_screens/home_screen_Student.dart';
 import 'package:hostelite/studentexitrecords.dart';
 import 'package:hostelite/students_complaint_list.dart';
 
@@ -35,7 +35,7 @@ class _StudentEntryRecordListState extends State<StudentEntryRecordList> {
       setState(() {
         my_entries = value.docs;
       });
-      my_entries.sort((b,a)=> a["time"].compareTo(b["time"]));
+      my_entries.sort((b, a) => a["time"].compareTo(b["time"]));
     });
   }
